fix: fix folder structure
